@@ -4,7 +4,8 @@ version: 1.0.0
 publish_to: none
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: '>=2.17.1 <3.0.0'
+  flutter: ">=3.0.1"
 
 dependencies:
   flutter:
@@ -15,8 +16,13 @@ dependencies:
     path: ../
 
 dev_dependencies:
+  flutter_driver:
+    sdk: flutter
+
   flutter_test:
     sdk: flutter
+
+  test:
 
 flutter:
   uses-material-design: true
