@@ -1,52 +1,52 @@
 name: upgrader
 description: Flutter package for prompting users to upgrade when there is a newer version of the app in the store.
-version: 4.6.1
+version: 6.0.1
 homepage: https://github.com/larryaasen/upgrader
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=1.12.0"
+  sdk: '>=2.18.6 <3.0.0'
+  flutter: ">=3.3.10"
 
 dependencies:
   flutter:
     sdk: flutter
 
   # From fluttercommunity.dev: Get current device information from within the Flutter application.
-  device_info_plus: ">=3.2.0 <4.2.0"
+  device_info_plus: ^8.0.0
+
+  # Pure Dart library for HTML5 parsing
+  html: ^0.15.1
 
   # For app_theme.dart
   google_fonts: ^2.3.1
 
   # From Dart Team: A composable, Future-based library for making HTTP requests.
-  http: ^0.13.0
+  http: ^0.13.5
 
   # From dart.dev: Platform independent access to information about the current operating system.
-  os_detect: ^2.0.0
+  os_detect: ^2.0.1
 
   # From Flutter Community: Provides an API for querying information about an application package.
-  package_info_plus: ^1.4.2
+  package_info_plus: ^3.0.1
 
   # From Flutter Team: Wraps NSUserDefaults (on iOS) and SharedPreferences (on Android).
-  shared_preferences: ^2.0.3
+  shared_preferences: ^2.0.15
 
   # From Flutter Team: A Flutter plugin for launching a URL in the mobile platform.
-  url_launcher: ^6.1.0
+  url_launcher: ^6.1.6
 
   # A dart library for comparing and incrementing version numbers with Semantic Versioning.
-  version: ">=2.0.0 <3.1.0"
+  version: ^3.0.2
 
   # A dart library for parsing, traversing, querying, transforming and building XML documents.
-  xml: ">=5.0.2 <7.0.0"
-
-  # Pure Dart library for HTML5 parsing
-  html: ^0.15.0
+  xml: ^6.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
   # From Dart Team: Mock library for Dart inspired by Mockito.
-  mockito: ^5.0.0
-  flutter_lints: ^1.0.4
+  mockito: ^5.3.2
+  flutter_lints: ^2.0.1
 
 flutter:
