@@ -2,8 +2,8 @@
  * Copyright (c) 2020-2022 Larry Aasen. All rights reserved.
  */
 
-import 'package:flutter/material.dart';
 import 'package:flutter/foundation.dart' show SynchronousFuture;
+import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:upgrader/upgrader.dart';
 
@@ -21,7 +21,7 @@ void main() async {
 }
 
 class Demo extends StatelessWidget {
-  Demo({Key key}) : super(key: key);
+  Demo({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -38,6 +38,7 @@ class Demo extends StatelessWidget {
         const Locale('en', ''), // English, no country code
         const Locale('ar', ''), // Arabic, no country code
         const Locale('bn', ''), // Bengali, no country code
+        const Locale('da', ''), // Danish, no country code
         const Locale('es', ''), // Spanish, no country code
         const Locale('fa', ''), // Persian, no country code
         const Locale('fil', ''), // Filipino, no country code
@@ -99,7 +100,7 @@ class DemoLocalizations {
   final Locale locale;
 
   static DemoLocalizations of(BuildContext context) {
-    return Localizations.of<DemoLocalizations>(context, DemoLocalizations);
+    return Localizations.of<DemoLocalizations>(context, DemoLocalizations)!;
   }
 
   static final Map<String, Map<String, String>> _localizedValues = {
@@ -114,11 +115,11 @@ class DemoLocalizations {
   };
 
   String get checking {
-    return _localizedValues[locale.languageCode]['checking'];
+    return _localizedValues[locale.languageCode]!['checking']!;
   }
 
   String get title {
-    return _localizedValues[locale.languageCode]['title'];
+    return _localizedValues[locale.languageCode]!['title']!;
   }
 }
 
@@ -131,6 +132,7 @@ class DemoLocalizationsDelegate
         'en',
         'ar',
         'bn',
+        'da',
         'es',
         'fa',
         'fil',
@@ -179,11 +181,11 @@ class MyUpgraderMessages extends UpgraderMessages {
   @override
   String get buttonTitleIgnore => 'My Ignore 1';
 
-  MyUpgraderMessages({String code}) : super(code: code);
+  MyUpgraderMessages({String? code}) : super(code: code);
 
   /// Override the message function to provide your own language localization.
   @override
-  String message(UpgraderMessage messageKey) {
+  String? message(UpgraderMessage messageKey) {
     if (languageCode == 'es') {
       switch (messageKey) {
         case UpgraderMessage.body:
