@@ -40,36 +40,45 @@ void main() {
     expect(upgrader1 == upgrader2, isTrue);
   }, skip: false);
 
+  testWidgets('test Upgrader clearSavedSettings', (WidgetTester tester) async {
+    await Upgrader.clearSavedSettings();
+  }, skip: false);
+
   testWidgets('test Upgrader class', (WidgetTester tester) async {
-    final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader(platform: TargetPlatform.iOS, client: client);
+    await tester.runAsync(() async {
+      // test code here
+      final client = MockITunesSearchClient.setupMockClient();
+      final upgrader = Upgrader(
+          platform: TargetPlatform.iOS, client: client, debugLogging: true);
 
-    expect(tester.takeException(), null);
-    await tester.pumpAndSettle();
-    try {
+      expect(tester.takeException(), null);
+      await tester.pumpAndSettle();
+      try {
+        expect(upgrader.appName(), 'Upgrader');
+      } catch (e) {
+        expect(e, upgrader.notInitializedExceptionMessage);
+      }
+
+      upgrader.installPackageInfo(
+          packageInfo: PackageInfo(
+              appName: 'Upgrader',
+              packageName: 'com.larryaasen.upgrader',
+              version: '1.9.9',
+              buildNumber: '400'));
+
+      await upgrader.initialize();
+
+      // Calling initialize() a second time should do nothing
+      await upgrader.initialize();
+
       expect(upgrader.appName(), 'Upgrader');
-    } catch (e) {
-      expect(e, upgrader.notInitializedExceptionMessage);
-    }
+      expect(upgrader.currentAppStoreVersion(), '5.6');
+      expect(upgrader.currentInstalledVersion(), '1.9.9');
+      expect(upgrader.isUpdateAvailable(), true);
 
-    upgrader.installPackageInfo(
-        packageInfo: PackageInfo(
-            appName: 'Upgrader',
-            packageName: 'com.larryaasen.upgrader',
-            version: '1.9.9',
-            buildNumber: '400'));
-    await upgrader.initialize();
-
-    // Calling initialize() a second time should do nothing
-    await upgrader.initialize();
-
-    expect(upgrader.appName(), 'Upgrader');
-    expect(upgrader.currentAppStoreVersion(), '5.6');
-    expect(upgrader.currentInstalledVersion(), '1.9.9');
-    expect(upgrader.isUpdateAvailable(), true);
-
-    upgrader.installAppStoreVersion('1.2.3');
-    expect(upgrader.currentAppStoreVersion(), '1.2.3');
+      upgrader.installAppStoreVersion('1.2.3');
+      expect(upgrader.currentAppStoreVersion(), '1.2.3');
+    });
   }, skip: false);
 
   testWidgets('test installAppStoreListingURL', (WidgetTester tester) async {
@@ -92,7 +101,8 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '0.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     var called = false;
     var notCalled = true;
@@ -126,7 +136,10 @@ void main() {
     expect(upgrader.messages.buttonTitleUpdate, 'ccc');
     expect(upgrader.messages.releaseNotes, 'ddd');
 
-    await tester.pumpWidget(_MyWidget(upgrader: upgrader));
+    // await tester.runAsync(() async {
+    final GlobalKey globalKey = GlobalKey();
+    final myWidget = _MyWidget(key: globalKey, upgrader: upgrader);
+    await tester.pumpWidget(myWidget);
 
     expect(find.text('Upgrader test'), findsOneWidget);
     expect(find.text('Upgrading'), findsOneWidget);
@@ -156,6 +169,7 @@ void main() {
     expect(find.text(upgrader.messages.releaseNotes), findsNothing);
     expect(called, true);
     expect(notCalled, true);
+    // });
   }, skip: false);
 
   testWidgets('test UpgradeWidget Cupertino', (WidgetTester tester) async {
@@ -169,7 +183,8 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '0.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     var called = false;
     var notCalled = true;
@@ -231,6 +246,7 @@ void main() {
     expect(called, true);
     expect(notCalled, true);
   }, skip: false);
+
   testWidgets('test UpgradeWidget ignore', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
     final upgrader = Upgrader(platform: TargetPlatform.iOS, client: client);
@@ -241,7 +257,8 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '0.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     var called = false;
     var notCalled = true;
@@ -283,7 +300,8 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '0.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     var called = false;
     var notCalled = true;
@@ -325,7 +343,8 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '0.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     var called = false;
     upgrader.shouldPopScope = () {
@@ -359,7 +378,8 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '0.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     expect(upgrader.messages, isNotNull);
 
@@ -406,7 +426,8 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '0.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     var called = false;
     var notCalled = true;
@@ -449,7 +470,8 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '0.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     var called = false;
     var notCalled = true;
@@ -493,7 +515,8 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '0.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     expect(upgrader.isTooSoon(), false);
     upgrader.minAppVersion = '0.5.0';
@@ -535,7 +558,8 @@ void main() {
             packageName: 'com.testing.test2',
             version: '2.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     expect(upgrader.belowMinAppVersion(), true);
     expect(upgrader.minAppVersion, '4.5.6');
@@ -555,7 +579,8 @@ void main() {
             packageName: 'com.larryaasen.upgrader',
             version: '2.9.9',
             buildNumber: '400'));
-    await upgrader.initialize();
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     expect(upgrader.belowMinAppVersion(), true);
     expect(upgrader.minAppVersion, '4.5.6');
@@ -567,7 +592,8 @@ void main() {
         platform: TargetPlatform.iOS,
         client: client,
         debugLogging: true,
-        countryCode: 'IT');
+        countryCode: 'IT',
+        languageCode: 'en');
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -575,7 +601,8 @@ void main() {
             packageName: 'com.google.MyApp',
             version: '0.1.0',
             buildNumber: '1'));
-    await upgrader.initialize();
+    upgrader.initialize().then((value) {});
+    await tester.pumpAndSettle();
 
     var called = false;
     var notCalled = true;
@@ -655,7 +682,7 @@ void main() {
 
       final upgrader2 =
           Upgrader(durationUntilAlertAgain: const Duration(days: 10));
-      final _ = UpgradeCard(upgrader: upgrader2);
+      UpgradeCard(upgrader: upgrader2);
       expect(upgrader2.durationUntilAlertAgain, const Duration(days: 10));
     }, skip: false);
 
@@ -799,6 +826,7 @@ void main() {
     verifyMessages(UpgraderMessages(code: 'en'), 'en');
     verifyMessages(UpgraderMessages(code: 'ar'), 'ar');
     verifyMessages(UpgraderMessages(code: 'bn'), 'bn');
+    verifyMessages(UpgraderMessages(code: 'da'), 'da');
     verifyMessages(UpgraderMessages(code: 'es'), 'es');
     verifyMessages(UpgraderMessages(code: 'fa'), 'fa');
     verifyMessages(UpgraderMessages(code: 'fil'), 'fil');
